pyproject.toml
[tool.poetry]
name = "mongolian"
version = "0.1.0"
description = ""
authors = ["Liang Hai <lianghai@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.3.1"
ufoLib2 = "^0.10.0"
tptqutils = { path = "../tptq/tptqutils/", develop = true }
tptqscripttools = { path = "../tptq/script-tools/", develop = true }
fontmake = "^2.3.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
